requirements.txt
# Required Dependencies
Flask==3.1.0
SQLAlchemy==2.0.38
Pydantic==2.10.6
python-keycloak==5.3.0
requests==2.32.3
grpcio==1.69.0
grpcio-tools==1.69.0
dependency-injector==4.45.0
pytest==8.3.4
pytest-mock==3.14.0

# Optional Dependencies
alembic==1.14.1
Flask-Migrate==4.1.0
loguru==0.7.3
Flask-CORS==5.0.0
python-dotenv==1.0.1
httpx==0.28.1
flask-swagger-ui==4.11.1
flask-restx==1.3.0
prometheus-flask-exporter==0.23.1